@@ -5,26 +5,44 @@ import android.net.Uri
 import android.os.Bundle
 import android.provider.MediaStore
 import android.widget.Toast
+import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
 import com.mediapicker.gallery.Gallery
 import com.mediapicker.gallery.GalleryConfig
 import com.mediapicker.gallery.domain.contract.IGalleryCommunicator
-import com.mediapicker.gallery.domain.entity.*
+import com.mediapicker.gallery.domain.entity.CarousalConfig
+import com.mediapicker.gallery.domain.entity.PhotoFile
+import com.mediapicker.gallery.domain.entity.PhotoTag
+import com.mediapicker.gallery.domain.entity.Rule
+import com.mediapicker.gallery.domain.entity.Validation
 import com.mediapicker.gallery.presentation.fragments.HomeFragment
 import com.mediapicker.gallery.presentation.utils.DefaultPage
 import com.mediapicker.gallery.presentation.viewmodels.VideoFile
-import kotlinx.android.synthetic.main.activity_main.*
+import com.mediapicker.sample.databinding.ActivityMainBinding
 import java.io.File
 
+private const val REQUEST_VIDEO_CAPTURE: Int = 1000
 
 class MainActivity : AppCompatActivity() {
 
-    private val REQUEST_VIDEO_CAPTURE: Int = 1000
     private var fragment: HomeFragment? = null
+
+    private var videoCaptureLauncher = registerForActivityResult(
+        ActivityResultContracts.StartActivityForResult()
+    ) { result ->
+        if (result.resultCode == RESULT_OK) {
+            val videoUri: Uri? = result.data?.data
+            fragment?.reloadMedia()
+        }
+    }
+
+    private val activityMainBinding: ActivityMainBinding by lazy {
+        ActivityMainBinding.inflate(layoutInflater)
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(activityMainBinding.root)
         setUpGallery()
         showStepFragment()
     }
@@ -36,15 +54,18 @@ class MainActivity : AppCompatActivity() {
             .build()
     }
 
-    private fun setUpGallery(){
-        val galleryConfig = GalleryConfig.GalleryConfigBuilder(applicationContext, BuildConfig.APPLICATION_ID + ".provider", MyClientGalleryCommunicator())
+    private fun setUpGallery() {
+        val galleryConfig = GalleryConfig.GalleryConfigBuilder(
+            BuildConfig.APPLICATION_ID + ".provider",
+            MyClientGalleryCommunicator()
+        )
             .useMyPhotoCamera(true)
             .useMyVideoCamera(false)
             .needToShowPreviewCarousal(CarousalConfig(true, 0, true, 0))
-            .mediaScanningCriteria(GalleryConfig.MediaScanningCriteria("",""))
+            .mediaScanningCriteria(GalleryConfig.MediaScanningCriteria("", ""))
             .typeOfMediaSupported(GalleryConfig.MediaType.PhotoWithFolderAndVideo)
             .validation(getValidation())
-            .photoTag( PhotoTag(true,"RC photo"))
+            .photoTag(PhotoTag(true, "RC photo"))
             .build()
         Gallery.init(galleryConfig)
     }
@@ -52,23 +73,26 @@ class MainActivity : AppCompatActivity() {
     private fun attachGalleryFragment() {
         try {
             val transaction = supportFragmentManager.beginTransaction()
-            val  photos =  SelectedItemHolder.listOfSelectedPhotos
-            fragment = DemoHomeFragment.getInstance(photos,
+            val photos = SelectedItemHolder.listOfSelectedPhotos
+            fragment = HomeFragment.getInstance(
+                photos,
                 SelectedItemHolder.listOfSelectedVideos,
                 defaultPageType = DefaultPage.PhotoPage
             )
-            transaction.replace(container.id, fragment!!, fragment!!::class.java.simpleName)
-            transaction.addToBackStack(fragment!!.javaClass.name)
-            transaction.commitAllowingStateLoss()
+            fragment?.let {
+                transaction.replace(activityMainBinding.container.id, it, it::class.java.simpleName)
+                transaction.addToBackStack(it.javaClass.name)
+                transaction.commitAllowingStateLoss()
+            }
         } catch (ex: Exception) {
             ex.printStackTrace()
         }
     }
 
     fun jumpToGallery() {
-       /* startActivity(GalleryActivity.getGalleryActivityIntent(SelectedItemHolder.listOfSelectedPhotos,
-            SelectedItemHolder.listOfSelectedVideos,
-            defaultPageType = DefaultPage.PhotoPage,context = baseContext))*/
+        /* startActivity(GalleryActivity.getGalleryActivityIntent(SelectedItemHolder.listOfSelectedPhotos,
+             SelectedItemHolder.listOfSelectedVideos,
+             defaultPageType = DefaultPage.PhotoPage,context = baseContext))*/
         attachGalleryFragment()
     }
 
@@ -76,7 +100,11 @@ class MainActivity : AppCompatActivity() {
         try {
             val transaction = supportFragmentManager.beginTransaction()
             val fragment = StepFragment()
-            transaction.replace(container.id, fragment, fragment::class.java.simpleName)
+            transaction.replace(
+                activityMainBinding.container.id,
+                fragment,
+                fragment::class.java.simpleName
+            )
             transaction.commitAllowingStateLoss()
         } catch (ex: Exception) {
             ex.printStackTrace()
@@ -89,7 +117,10 @@ class MainActivity : AppCompatActivity() {
 
         }
 
-        override fun actionButtonClick(listOfSelectedPhotos: List<PhotoFile>, listofSelectedVideos: List<VideoFile>) {
+        override fun actionButtonClick(
+            listOfSelectedPhotos: List<PhotoFile>,
+            listofSelectedVideos: List<VideoFile>
+        ) {
             SelectedItemHolder.listOfSelectedPhotos = listOfSelectedPhotos.toMutableList()
             SelectedItemHolder.listOfSelectedVideos = listofSelectedVideos
             showStepFragment()
@@ -116,35 +147,31 @@ class MainActivity : AppCompatActivity() {
         }
 
         private fun showMessage(msg: String) {
-             Toast.makeText(applicationContext,msg,Toast.LENGTH_LONG).show()
+            Toast.makeText(applicationContext, msg, Toast.LENGTH_LONG).show()
 
         }
 
         override fun onPermissionDenied() {
-            Toast.makeText(applicationContext,"Permission denied :(",Toast.LENGTH_LONG).show()
+            Toast.makeText(applicationContext, "Permission denied :(", Toast.LENGTH_LONG).show()
         }
 
         override fun onNeverAskPermissionAgain() {
-            Toast.makeText(applicationContext,"Permission denied :(",Toast.LENGTH_LONG).show()
+            Toast.makeText(applicationContext, "Permission denied :(", Toast.LENGTH_LONG).show()
+        }
+
+//        override fun onShowPermissionRationale(permissionRequest: PermissionRequestWrapper) {
+//            Toast.makeText(applicationContext, "Permission show rationale :|", Toast.LENGTH_LONG)
+//                .show()
+//        }
+
+        override fun onStepValidate(isValid: Boolean) {
         }
     }
 
 
     private fun dispatchTakeVideoIntent() {
-        Intent(MediaStore.ACTION_VIDEO_CAPTURE).also { takeVideoIntent ->
-            takeVideoIntent.resolveActivity(packageManager)?.also {
-                startActivityForResult(takeVideoIntent, REQUEST_VIDEO_CAPTURE)
-            }
-        }
-    }
-
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        super.onActivityResult(requestCode, resultCode, data)
-        if (requestCode == REQUEST_VIDEO_CAPTURE && resultCode == RESULT_OK) {
-            val videoUri: Uri? = intent.data
-            Toast.makeText(baseContext, "Recorded ", Toast.LENGTH_LONG).show()
-            fragment?.reloadMedia()
-        }
+        val videoIntent = Intent(MediaStore.ACTION_VIDEO_CAPTURE)
+        videoCaptureLauncher.launch(videoIntent)
     }
 
 }
@@ -160,11 +187,11 @@ object SelectedItemHolder {
 //        builder.fullPhotoUrl("https://www.hackingwithswift.com/uploads/matrix.jpg")
 //        this.add(builder.build())
 
-     /*   val builder1 = PhotoFile.Builder()
-        builder1.apolloKey = "11112"
-        builder1.imageId = 20
-        builder1.fullPhotoUrl("https://www.hackingwithswift.com/uploads/matrix.jpg")
-        this.add(builder1.build())*/
+    /*   val builder1 = PhotoFile.Builder()
+       builder1.apolloKey = "11112"
+       builder1.imageId = 20
+       builder1.fullPhotoUrl("https://www.hackingwithswift.com/uploads/matrix.jpg")
+       this.add(builder1.build())*/
 //    }
     var listOfSelectedVideos = emptyList<VideoFile>()
 }
