@@ -1,19 +1,17 @@
 package com.mediapicker.gallery.data.repositories
 
-import android.app.Application
-import android.content.Context
+import android.content.ContentResolver
 import android.database.Cursor
 import android.provider.MediaStore
-import android.util.Log
 import android.webkit.MimeTypeMap
 import com.mediapicker.gallery.domain.entity.PhotoAlbum
 import com.mediapicker.gallery.domain.entity.PhotoFile
 import com.mediapicker.gallery.domain.repositories.GalleryRepository
 
-open class GalleryService(private val applicationContext: Context) : GalleryRepository {
+open class GalleryService(private val contentResolver: ContentResolver) : GalleryRepository {
 
     companion object {
-        fun getInstance(context: Application) = GalleryService(context)
+        //        fun getInstance(context: Application) = GalleryService(context)
         const val COL_FULL_PHOTO_URL = "fullPhotoUrl"
     }
 
@@ -30,7 +28,7 @@ open class GalleryService(private val applicationContext: Context) : GalleryRepo
         val mimeTypeGif = MimeTypeMap.getSingleton().getMimeTypeFromExtension("gif")
         val selectionTypeGifArgs = arrayOf(mimeTypeGif)
         val cursor = MediaStore.Images.Media.query(
-            applicationContext.contentResolver,
+            contentResolver,
             MediaStore.Images.Media.EXTERNAL_CONTENT_URI, null, selection, selectionTypeGifArgs,
             MediaStore.Images.Media.DATE_ADDED + " DESC"
         )
@@ -70,8 +68,10 @@ open class GalleryService(private val applicationContext: Context) : GalleryRepo
     }
 
     private fun getPhoto(cursor: Cursor): PhotoFile {
-        val id = cursor.getLong(cursor.getColumnIndex(MediaStore.Images.Media._ID))
-        val path = cursor.getString(cursor.getColumnIndex(MediaStore.Images.Media.DATA))
+        val id = cursor.getLong(cursor.getColumnIndexOrThrow(MediaStore.Images.Media._ID))
+        val path = cursor.getString(cursor.getColumnIndexOrThrow(MediaStore.Images.Media.DATA))
+        val mimeType =
+            cursor.getString(cursor.getColumnIndexOrThrow(MediaStore.Images.Media.MIME_TYPE))
         val col = cursor.getColumnIndex(COL_FULL_PHOTO_URL)
         var fullPhotoUrl = ""
         if (col != -1) {
@@ -83,6 +83,7 @@ open class GalleryService(private val applicationContext: Context) : GalleryRepo
             .smallPhotoUrl("")
             .fullPhotoUrl(fullPhotoUrl)
             .photoBackendId(0L)
+            .mimeType()
             .build()
     }
 
