@@ -10,10 +10,10 @@ class CameraItem : IGalleryItem, Serializable {
     var albumEntries: List<PhotoFile> = ArrayList()
 
 
-    override fun equals(o: Any?): Boolean {
-        if (this === o) return true
-        if (o == null || javaClass != o.javaClass) return false
-        val that = o as CameraItem?
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (other == null || javaClass != other.javaClass) return false
+        val that = other as CameraItem?
         return if (this.albumId == null) {
             that!!.albumId == null
         } else this.albumId == that!!.albumId
