@@ -1,3 +1,10 @@
 package com.mediapicker.gallery.domain.entity
 
-data class GalleryLabels(val homeTitle: String = "", val homeAction: String = "", val galleryFolderAction: String = "")
+import android.view.Gravity
+
+data class GalleryLabels(
+    val homeTitle: String? = "",
+    val homeAction: String? = "",
+    val galleryFolderAction: String? = "",
+    val titleAlignment: Int = Gravity.START or Gravity.CENTER_VERTICAL
+)
