@@ -26,10 +26,10 @@ class PhotoAlbum(var albumId: String?, var name: String?) : IGalleryItem, Serial
     }
 
 
-    override fun equals(o: Any?): Boolean {
-        if (this === o) return true
-        if (o == null || javaClass != o.javaClass) return false
-        val that = o as PhotoAlbum?
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (other == null || javaClass != other.javaClass) return false
+        val that = other as PhotoAlbum?
         return if (this.albumId == null) {
             that!!.albumId == null
         } else this.albumId == that!!.albumId
