@@ -1,4 +1,3 @@
 package com.mediapicker.gallery.domain.entity
 
-data class PhotoTag(val shouldShowPhotoTag:Boolean=false, var photoTagText:String="") {
-}
+data class PhotoTag(val shouldShowPhotoTag: Boolean = false, var photoTagText: String = "")
