@@ -2,4 +2,4 @@ package com.mediapicker.gallery.domain.entity
 
 import java.io.Serializable
 
-data class UploadedPhoto constructor(val id : String,val url : String) : Serializable
+data class UploadedPhoto(val id: String, val url: String) : Serializable
