@@ -1,27 +1,25 @@
 package com.mediapicker.gallery.presentation.activity
 
 import android.os.Bundle
-import androidx.annotation.LayoutRes
 import androidx.appcompat.app.AppCompatActivity
-import com.mediapicker.gallery.R
+import com.mediapicker.gallery.databinding.OssBaseFragmentActivityBinding
 import com.mediapicker.gallery.presentation.fragments.BaseFragment
-import kotlinx.android.synthetic.main.oss_base_fragment_activity.*
 
 abstract class BaseFragmentActivity : AppCompatActivity() {
 
+    private val binding: OssBaseFragmentActivityBinding by lazy {
+        OssBaseFragmentActivityBinding.inflate(layoutInflater)
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(getLayout())
+        setContentView(binding.root)
     }
-
-    @LayoutRes
-    fun getLayout() = R.layout.oss_base_fragment_activity
 
     protected fun setFragment(fragment: BaseFragment, addToBackStack: Boolean = true) {
         try {
             val transaction = supportFragmentManager.beginTransaction()
-            transaction.add(container.id, fragment, fragment.javaClass.name)
+            transaction.add(binding.container.id, fragment, fragment.javaClass.name)
             if (addToBackStack)
                 transaction.addToBackStack(fragment.javaClass.name)
             transaction.commitAllowingStateLoss()
