@@ -3,6 +3,7 @@ package com.mediapicker.gallery.presentation.activity
 import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
+import androidx.activity.OnBackPressedCallback
 import androidx.fragment.app.Fragment
 import com.mediapicker.gallery.domain.entity.PhotoAlbum
 import com.mediapicker.gallery.domain.entity.PhotoFile
@@ -16,24 +17,18 @@ import com.mediapicker.gallery.presentation.utils.Constants.PHOTO_SELECTION_REQU
 
 class FolderViewActivity : BaseFragmentActivity(), GalleryActionListener {
 
-    companion object {
-        fun startActivityForResult(fragment: Fragment, currentSelectPhotos: LinkedHashSet<PhotoFile>) {
-            fragment.startActivityForResult(Intent(fragment.activity, FolderViewActivity::class.java).apply {
-                this.putExtra(EXTRA_SELECTED_PHOTO, currentSelectPhotos)
-            }, PHOTO_SELECTION_REQUEST_CODE)
-        }
-    }
-
     private var currentSelectedPhotos = LinkedHashSet<PhotoFile>()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setCurrentSelectedPhotos()
         setFragment(FolderViewFragment.getInstance())
+        onBackPressedDispatcher.addCallback(this, onBackPressedCallback)
     }
 
     private fun setCurrentSelectedPhotos() {
-        currentSelectedPhotos = intent.getSerializableExtra(EXTRA_SELECTED_PHOTO) as LinkedHashSet<PhotoFile>
+        currentSelectedPhotos =
+            intent.getSerializableExtra(EXTRA_SELECTED_PHOTO) as LinkedHashSet<PhotoFile>
     }
 
     override fun moveToPhotoGrid(photoAlbum: PhotoAlbum) {
@@ -58,10 +53,12 @@ class FolderViewActivity : BaseFragmentActivity(), GalleryActionListener {
 
     override fun onActionClicked(shouldThrowResult: Boolean) {
         if (shouldThrowResult) {
-            setResult(Activity.RESULT_OK, Intent().apply { this.putExtra(EXTRA_SELECTED_PHOTO, currentSelectedPhotos) })
+            setResult(
+                Activity.RESULT_OK,
+                Intent().apply { this.putExtra(EXTRA_SELECTED_PHOTO, currentSelectedPhotos) })
             finish()
         } else {
-            onBackPressed()
+            onBackPressedDispatcher.onBackPressed()
         }
     }
 
@@ -69,12 +66,27 @@ class FolderViewActivity : BaseFragmentActivity(), GalleryActionListener {
         // showCloseButton()
     }
 
-    override fun onBackPressed() {
-        val fragments = supportFragmentManager.backStackEntryCount
-        when {
-            fragments == 1 -> finish()
-            fragments > 1 -> supportFragmentManager.popBackStack()
-            else -> super.onBackPressed()
+//    override fun onBackPressed() {
+//        val fragments = supportFragmentManager.backStackEntryCount
+//        when {
+//            fragments == 1 -> finish()
+//            fragments > 1 -> supportFragmentManager.popBackStack()
+//            else -> super.onBackPressed()
+//        }
+//    }
+
+    private val onBackPressedCallback = object : OnBackPressedCallback(true) {
+        override fun handleOnBackPressed() {
+            val fragments = supportFragmentManager.backStackEntryCount
+            when {
+                fragments == 1 -> finish()
+                fragments > 1 -> supportFragmentManager.popBackStack()
+                else -> {
+                    isEnabled = false
+                    onBackPressedDispatcher.onBackPressed()
+                    isEnabled = true
+                }
+            }
         }
     }
 }
