@@ -11,33 +11,57 @@ import com.bumptech.glide.Glide
 import com.bumptech.glide.request.RequestOptions
 import com.mediapicker.gallery.Gallery
 import com.mediapicker.gallery.R
+import com.mediapicker.gallery.databinding.OssItemFolderSelectionBinding
 import com.mediapicker.gallery.domain.contract.OnItemClickListener
 import com.mediapicker.gallery.domain.entity.PhotoAlbum
 import com.mediapicker.gallery.domain.entity.PhotoFile
-import kotlinx.android.synthetic.main.oss_item_folder_selection.view.*
 import java.io.File
 
-open class GalleryFolderAdapter constructor(val context: Context, var listOfFolders: List<PhotoAlbum>, val onItemClickListener: OnItemClickListener<PhotoAlbum>?=null) : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
+open class GalleryFolderAdapter(
+    val context: Context,
+    var listOfFolders: List<PhotoAlbum>,
+    val onItemClickListener: OnItemClickListener<PhotoAlbum>? = null
+) : RecyclerView.Adapter<FolderViewHolder>() {
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
-        val view = LayoutInflater.from(parent.context).inflate(R.layout.oss_item_folder_selection, parent, false)
-        return FolderViewHolder(view)
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): FolderViewHolder {
+        val binding: OssItemFolderSelectionBinding = OssItemFolderSelectionBinding.inflate(
+            LayoutInflater.from(parent.context),
+            parent,
+            false
+        )
+        return FolderViewHolder(binding, onItemClickListener)
     }
 
     override fun getItemCount() = listOfFolders.size
 
 
-    override fun onBindViewHolder(viewHolder: RecyclerView.ViewHolder, position: Int) {
-        val vH = viewHolder as FolderViewHolder
-        vH.root.setOnClickListener { v -> onItemClickListener?.onListItemClick(listOfFolders[position]) }
-        vH.root.folderName.text = listOfFolders[position].name
-        vH.root.folderName.isAllCaps = Gallery.galleryConfig.textAllCaps
-        val album = listOfFolders[position]
+    override fun onBindViewHolder(viewHolder: FolderViewHolder, position: Int) {
+        viewHolder.bind(listOfFolders[position])
+    }
+}
+
+
+class FolderViewHolder(
+    private val binding: OssItemFolderSelectionBinding,
+    private val onItemClickListener: OnItemClickListener<PhotoAlbum>?
+) :
+    RecyclerView.ViewHolder(binding.root) {
+    init {
+        binding.img.setImageResource(R.drawable.oss_media_ic_folder_icon)
+    }
+
+    fun bind(album: PhotoAlbum) {
+        binding.root.setOnClickListener { onItemClickListener?.onListItemClick(album) }
+        binding.folderName.text = album.name
+        binding.folderName.isAllCaps = Gallery.galleryConfig.textAllCaps
         if (album.hasPhotos()) {
-            vH.root.backgroundImage.visibility = View.VISIBLE
-            loadImageIntoView(album.firstPhoto as PhotoFile, vH.root.backgroundImage)  //todo shalini: change the type inference
+            binding.backgroundImage.visibility = View.VISIBLE
+            loadImageIntoView(
+                album.firstPhoto as PhotoFile,
+                binding.backgroundImage
+            )
         } else {
-            vH.root.backgroundImage.visibility = View.GONE
+            binding.backgroundImage.visibility = View.GONE
         }
     }
 
@@ -45,25 +69,19 @@ open class GalleryFolderAdapter constructor(val context: Context, var listOfFold
         val options = RequestOptions()
         if (photo.isAlreadyUploaded) {
             photo.path?.let {
-                Glide.with(imageView.context).load(photo.fullPhotoUrl).apply(options).into(imageView)
+                Glide.with(imageView.context).load(photo.fullPhotoUrl).apply(options)
+                    .into(imageView)
             }
         } else {
             if (photo.existsPhoto()) {
                 photo.path?.let {
-                    Glide.with(imageView.context).load(Uri.fromFile(File(it))).apply(options).into(imageView)
+                    Glide.with(imageView.context).load(Uri.fromFile(File(it))).apply(options)
+                        .into(imageView)
                 }
             } else {
                 photo.toString()
             }
         }
-    }
-
-}
-
-
-internal class FolderViewHolder(var root: View) : RecyclerView.ViewHolder(root) {
-    init {
-        root.img.setImageResource(R.drawable.oss_media_ic_folder_icon)
     }
 }
 
