@@ -5,6 +5,7 @@ import com.mediapicker.gallery.domain.entity.PhotoFile
 interface CarousalActionListener {
     fun onItemClicked(photoFile: PhotoFile, isSelected: Boolean)
     fun onGalleryFolderSelected()
-    fun onGalleryImagePreview()
+    fun onGalleryImagePreview(index: Int, listSize: Int)
     fun onGalleryImagePreviewChanged()
+    fun onGalleryImagePreviewClosed(index: Int, size: Int)
 }
