@@ -10,10 +10,17 @@ import com.google.android.material.appbar.AppBarLayout
 class FlingBehavior : AppBarLayout.Behavior {
     private var isPositive = false
 
-    constructor() {}
-    constructor(context: Context?, attrs: AttributeSet?) : super(context, attrs) {}
+    constructor()
+    constructor(context: Context?, attrs: AttributeSet?) : super(context, attrs)
 
-    override fun onNestedFling(coordinatorLayout: CoordinatorLayout, child: AppBarLayout, target: View, velocityX: Float, velocityY: Float, consumed: Boolean): Boolean {
+    override fun onNestedFling(
+        coordinatorLayout: CoordinatorLayout,
+        child: AppBarLayout,
+        target: View,
+        velocityX: Float,
+        velocityY: Float,
+        consumed: Boolean
+    ): Boolean {
         var velocityY = velocityY
         var consumed = consumed
         if (velocityY > 0 && !isPositive || velocityY < 0 && isPositive) {
@@ -27,8 +34,21 @@ class FlingBehavior : AppBarLayout.Behavior {
         return super.onNestedFling(coordinatorLayout, child, target, velocityX, velocityY, consumed)
     }
 
-    override fun onNestedPreScroll(coordinatorLayout: CoordinatorLayout, child: AppBarLayout, target: View, dx: Int, dy: Int, consumed: IntArray) {
-        super.onNestedPreScroll(coordinatorLayout, child, target, dx, dy, consumed)
+//    override fun onNestedPreScroll(coordinatorLayout: CoordinatorLayout, child: AppBarLayout, target: View, dx: Int, dy: Int, consumed: IntArray) {
+//        super.onNestedPreScroll(coordinatorLayout, child, target, dx, dy, consumed)
+//        isPositive = dy > 0
+//    }
+
+    override fun onNestedPreScroll(
+        coordinatorLayout: CoordinatorLayout,
+        child: AppBarLayout,
+        target: View,
+        dx: Int,
+        dy: Int,
+        consumed: IntArray,
+        type: Int
+    ) {
+        super.onNestedPreScroll(coordinatorLayout, child, target, dx, dy, consumed, type)
         isPositive = dy > 0
     }
 
