@@ -6,13 +6,13 @@ import android.view.MotionEvent
 import androidx.viewpager.widget.ViewPager
 
 class MediaGalleryViewPager : ViewPager {
-    constructor(context: Context?) : super(context!!) {}
-    constructor(context: Context?, attrs: AttributeSet?) : super(context!!, attrs) {}
+    constructor(context: Context?) : super(context!!)
+    constructor(context: Context?, attrs: AttributeSet?) : super(context!!, attrs)
 
     override fun onInterceptTouchEvent(ev: MotionEvent): Boolean {
         try {
             return super.onInterceptTouchEvent(ev)
-        } catch (e: IllegalArgumentException) {
+        } catch (_: IllegalArgumentException) {
         }
         return false
     }
