@@ -7,46 +7,60 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.annotation.CallSuper
 import androidx.annotation.LayoutRes
+import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import com.mediapicker.gallery.R
+import com.mediapicker.gallery.databinding.OssFragmentBaseBinding
 import com.mediapicker.gallery.domain.entity.PhotoFile
 import com.mediapicker.gallery.presentation.viewmodels.VideoFile
-import kotlinx.android.synthetic.main.oss_custom_toolbar.*
-import kotlinx.android.synthetic.main.oss_fragment_base.*
-import kotlinx.android.synthetic.main.oss_fragment_base.view.*
 
-open abstract class BaseFragment : Fragment() {
-
-    companion object{
+abstract class BaseFragment : Fragment() {
+    //    protected var binding: T? = null
+//    <T : ViewBinding>(contentLayoutId: Int)
+    companion object {
         const val EXTRA_SELECTED_PHOTOS = "selected_photos"
         const val EXTRA_SELECTED_VIDEOS = "selected_videos"
         const val EXTRA_DEFAULT_PAGE = "extra_default_page"
     }
 
+    lateinit var childView: View
+
     @Suppress("UNCHECKED_CAST")
-    protected fun getPhotosFromArguments() : List<PhotoFile>{
+    protected fun getPhotosFromArguments(): List<PhotoFile> {
         this.arguments?.let {
-            if(it.containsKey(EXTRA_SELECTED_PHOTOS)){
+            if (it.containsKey(EXTRA_SELECTED_PHOTOS)) {
                 return it.getSerializable(EXTRA_SELECTED_PHOTOS) as List<PhotoFile>
             }
         }
         return emptyList()
     }
 
+//    protected abstract fun inflateBiding(inflater: LayoutInflater, container: ViewGroup?): T
+
     @Suppress("UNCHECKED_CAST")
-    protected fun getVideosFromArguments() : List<VideoFile>{
-       this.arguments?.let {
-            if(it.containsKey(EXTRA_SELECTED_VIDEOS)){
+    protected fun getVideosFromArguments(): List<VideoFile> {
+        this.arguments?.let {
+            if (it.containsKey(EXTRA_SELECTED_VIDEOS)) {
                 return it.getSerializable(EXTRA_SELECTED_VIDEOS) as List<VideoFile>
             }
         }
-        return  emptyList()
+        return emptyList()
     }
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.oss_fragment_base, container, false).apply {
-            baseContainer.addView(inflater.inflate(getLayoutId(),null))
+    var ossFragmentBaseBinding: OssFragmentBaseBinding? = null
+
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        ossFragmentBaseBinding = OssFragmentBaseBinding.inflate(inflater, container, false).apply {
+            childView = inflater.inflate(getLayoutId(), null)
+
+            baseContainer.addView(childView)
         }
+        return ossFragmentBaseBinding?.root
     }
 
     @LayoutRes
@@ -61,55 +75,65 @@ open abstract class BaseFragment : Fragment() {
 
     @CallSuper
     private fun setToolbar() {
-        toolbarTitle.text = getScreenTitle()
-        toolbarTitle.setTextColor(context!!.resources!!.getColor(R.color.oss_toolbar_text))
-        if(setHomeAsUp()){
-            toolbarBackButton.visibility = View.VISIBLE
-            toolbarBackButton.setImageResource(getHomeAsUpIcon())
-        }else{
-            toolbarBackButton.visibility = View.GONE
+        ossFragmentBaseBinding?.ossCustomTool?.toolbarTitle?.text = getScreenTitle()
+        context?.let {
+            ossFragmentBaseBinding?.ossCustomTool?.toolbarTitle?.setTextColor(
+                ContextCompat.getColor(
+                    it,
+                    R.color.oss_toolbar_text
+                )
+            )
         }
-        if(shouldHideToolBar()){
-            toolbarView.visibility = View.GONE
+        if (setHomeAsUp()) {
+            ossFragmentBaseBinding?.ossCustomTool?.toolbarBackButton?.visibility = View.VISIBLE
+            ossFragmentBaseBinding?.ossCustomTool?.toolbarBackButton?.setImageResource(
+                getHomeAsUpIcon()
+            )
+        } else {
+            ossFragmentBaseBinding?.ossCustomTool?.toolbarBackButton?.visibility = View.GONE
         }
-        toolbarBackButton.setOnClickListener { onBackPressed() }
+        if (shouldHideToolBar()) {
+            ossFragmentBaseBinding?.ossCustomTool?.root?.visibility = View.GONE
+        }
+        ossFragmentBaseBinding?.ossCustomTool?.toolbarBackButton?.setOnClickListener { onBackPressed() }
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        return if(item.itemId == android.R.id.home){
+        return if (item.itemId == android.R.id.home) {
             onBackPressed()
             return true
-        }else{
+        } else {
             super.onOptionsItemSelected(item)
         }
-
     }
 
     protected fun showToolbar() {
-        toolbarView.visibility = View.VISIBLE
+        ossFragmentBaseBinding?.ossCustomTool?.root?.visibility = View.VISIBLE
     }
 
     protected fun hideToolbar() {
-        toolbarView.visibility = View.GONE
+        ossFragmentBaseBinding?.ossCustomTool?.root?.visibility = View.GONE
     }
 
     abstract fun onBackPressed()
 
-    open fun getHomeAsUpIcon()  = R.drawable.oss_media_ic_back
+    open fun getHomeAsUpIcon() = R.drawable.oss_media_ic_back
 
-    open fun setHomeAsUp()  = false
+    open fun setHomeAsUp() = false
 
-    abstract fun getScreenTitle() : String
+    abstract fun getScreenTitle(): String
 
-    open fun shouldHideToolBar()  = false
+    open fun shouldHideToolBar() = false
 
     abstract fun setUpViews()
 
     @CallSuper
-    open fun initViewModels(){}
+    open fun initViewModels() {
+    }
 
     override fun onDestroyView() {
         super.onDestroyView()
-        baseToolbar.visibility = View.VISIBLE
+        ossFragmentBaseBinding?.baseToolbar?.visibility = View.VISIBLE
+        ossFragmentBaseBinding = null
     }
 }
