@@ -1,29 +1,20 @@
 package com.mediapicker.gallery.presentation.fragments
 
-import android.app.Activity
+import android.content.Context
 import com.mediapicker.gallery.presentation.activity.GalleryActionListener
-import kotlinx.android.synthetic.main.oss_fragment_folder_view.*
 
 abstract class BaseGalleryViewFragment : BaseFragment() {
 
-
     protected var galleryActionListener: GalleryActionListener? = null
 
-    override fun onAttach(activity: Activity) {
-        super.onAttach(activity)
-        if (activity is GalleryActionListener) {
-            galleryActionListener = activity
+    override fun onAttach(context: Context) {
+        super.onAttach(context)
+        if (context is GalleryActionListener) {
+            galleryActionListener = context
         }
     }
 
-    override fun setUpViews() {
-        actionButton.setOnClickListener { onActionButtonClick() }
-    }
-
-    open fun onActionButtonClick() {}
-
     override fun onBackPressed() {
-        activity?.onBackPressed()
+        activity?.onBackPressedDispatcher?.onBackPressed()
     }
-
 }
