@@ -2,10 +2,12 @@ package com.mediapicker.gallery.presentation.fragments
 
 import android.annotation.SuppressLint
 import android.os.Bundle
+import android.widget.LinearLayout
 import androidx.recyclerview.widget.GridLayoutManager
 import com.google.android.material.snackbar.Snackbar
 import com.mediapicker.gallery.Gallery
 import com.mediapicker.gallery.R
+import com.mediapicker.gallery.databinding.OssFragmentFolderViewBinding
 import com.mediapicker.gallery.domain.entity.PhotoAlbum
 import com.mediapicker.gallery.domain.entity.PhotoFile
 import com.mediapicker.gallery.presentation.adapters.IGalleryItemClickListener
@@ -16,9 +18,6 @@ import com.mediapicker.gallery.presentation.utils.ItemDecorationAlbumColumns
 import com.mediapicker.gallery.presentation.utils.ValidatePhotos
 import com.mediapicker.gallery.presentation.utils.ValidationResult
 import com.mediapicker.gallery.utils.SnackbarUtils
-import kotlinx.android.synthetic.main.oss_custom_toolbar.*
-import kotlinx.android.synthetic.main.oss_fragment_carousal.*
-import kotlinx.android.synthetic.main.oss_fragment_folder_view.*
 
 const val COLUMNS_COUNT = 3
 
@@ -36,7 +35,12 @@ class GalleryPhotoViewFragment : BaseGalleryViewFragment() {
         arguments?.getSerializable(EXTRA_SELECTED_PHOTO) as LinkedHashSet<PhotoFile>
     }
 
-    private fun removePhotoFromSelection(photo: PhotoFile, position: Int) {
+    private val ossFragmentFolderView: OssFragmentFolderViewBinding? by lazy {
+        ossFragmentBaseBinding?.baseContainer?.findViewById<LinearLayout>(R.id.linear_layout_parent)
+            ?.let { OssFragmentFolderViewBinding.bind(it) }
+    }
+
+    private fun removePhotoFromSelection(photo: PhotoFile) {
         currentSelectedPhotos.removePhoto(photo)
         Gallery.carousalActionListener?.onItemClicked(photo, false)
         adapter.listCurrentPhotos = currentSelectedPhotos.toList()
@@ -52,50 +56,70 @@ class GalleryPhotoViewFragment : BaseGalleryViewFragment() {
     override fun getLayoutId() = R.layout.oss_fragment_folder_view
 
     override fun setUpViews() {
-        super.setUpViews()
+        ossFragmentFolderView?.actionButton?.setOnClickListener { onActionButtonClick() }
+
         photoAlbum.let { album ->
-            adapter = SelectPhotoImageAdapter(album.getAlbumEntries(), currentSelectedPhotos.toList(), galleryItemClickListener, fromGallery = false)
+            adapter = SelectPhotoImageAdapter(
+                album.getAlbumEntries(),
+                currentSelectedPhotos.toList(),
+                galleryItemClickListener,
+                fromGallery = false
+            )
         }
 
-        folderRV.apply {
-            this.addItemDecoration(ItemDecorationAlbumColumns(resources.getDimensionPixelSize(R.dimen.module_base), COLUMNS_COUNT))
+        ossFragmentFolderView?.folderRV?.apply {
+            this.addItemDecoration(
+                ItemDecorationAlbumColumns(
+                    resources.getDimensionPixelSize(R.dimen.module_base),
+                    COLUMNS_COUNT
+                )
+            )
             this.layoutManager = GridLayoutManager(activity, COLUMNS_COUNT)
             this.adapter = this@GalleryPhotoViewFragment.adapter
         }
 
-        if (Gallery.galleryConfig.galleryLabels.galleryFolderAction.isNotBlank()) {
-            actionButton.text = Gallery.galleryConfig.galleryLabels.galleryFolderAction
+        ossFragmentFolderView?.actionButton?.isSelected = true
+
+        if (Gallery.galleryConfig.galleryLabels.galleryFolderAction?.isNotBlank() == true) {
+            ossFragmentFolderView?.actionButton?.text =
+                Gallery.galleryConfig.galleryLabels.galleryFolderAction
         }
-        toolbarTitle.isAllCaps = Gallery.galleryConfig.textAllCaps
-        actionButton.isAllCaps = Gallery.galleryConfig.textAllCaps
+        ossFragmentBaseBinding?.ossCustomTool?.toolbarTitle?.isAllCaps =
+            Gallery.galleryConfig.textAllCaps
+        ossFragmentFolderView?.actionButton?.isAllCaps = Gallery.galleryConfig.textAllCaps
+
+//        baseBinding.customToolbar.apply {
+//            toolbarTitle.isAllCaps = Gallery.galleryConfig.textAllCaps
+//            toolbarTitle.gravity = Gallery.galleryConfig.galleryLabels.titleAlignment
+//        }
     }
 
     @SuppressLint("CheckResult")
     private fun handleItemClickListener(photo: PhotoFile, position: Int) {
         if (currentSelectedPhotos.containsPhoto(photo)) {
-            removePhotoFromSelection(photo, position)
+            removePhotoFromSelection(photo)
         } else {
-            validateNewPhoto(photo, position)
+            validateNewPhoto(photo)
         }
     }
 
-    private fun validateNewPhoto(photo: PhotoFile, position: Int) {
-        when(val validationResult = photoValidationAction.canAddThisToList(currentSelectedPhotos.size, photo)){
+    private fun validateNewPhoto(photo: PhotoFile) {
+        when (val validationResult =
+            photoValidationAction.canAddThisToList(currentSelectedPhotos.size, photo)) {
             is ValidationResult.Success -> {
                 galleryActionListener?.onPhotoSelected(photo)
                 Gallery.carousalActionListener?.onItemClicked(photo, true)
                 adapter.listCurrentPhotos = currentSelectedPhotos.toList()
                 adapter.notifyDataSetChanged()
             }
+
             is ValidationResult.Failure -> {
-                var msg = validationResult.msg
-                showError(msg)
+                showError(validationResult.msg)
             }
         }
     }
 
-    override fun onActionButtonClick() {
-        super.onActionButtonClick()
+    private fun onActionButtonClick() {
         galleryActionListener?.onActionClicked(false)
     }
 
@@ -103,8 +127,8 @@ class GalleryPhotoViewFragment : BaseGalleryViewFragment() {
 
     private val galleryItemClickListener = object : IGalleryItemClickListener {
 
-        override fun onPhotoItemClick(photo: PhotoFile, position: Int) {
-            handleItemClickListener(photo, position)
+        override fun onPhotoItemClick(photoFile: PhotoFile, position: Int) {
+            handleItemClickListener(photoFile, position)
         }
 
         override fun onFolderItemClick() {
@@ -117,7 +141,10 @@ class GalleryPhotoViewFragment : BaseGalleryViewFragment() {
     }
 
     companion object {
-        fun getInstance(photoAlbum: PhotoAlbum, currentSelectedPhotos: java.util.LinkedHashSet<PhotoFile>) = GalleryPhotoViewFragment().apply {
+        fun getInstance(
+            photoAlbum: PhotoAlbum,
+            currentSelectedPhotos: java.util.LinkedHashSet<PhotoFile>
+        ) = GalleryPhotoViewFragment().apply {
             arguments = Bundle().apply {
                 this.putSerializable(EXTRA_SELECTED_ALBUM, photoAlbum)
                 this.putSerializable(EXTRA_SELECTED_PHOTO, currentSelectedPhotos)
