@@ -4,15 +4,14 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.mediapicker.gallery.domain.entity.PhotoAlbum
 import com.mediapicker.gallery.domain.repositories.GalleryRepository
-import java.util.*
 
-class LoadAlbumViewModel constructor(private val galleryRepository: GalleryRepository) : ViewModel(){
+class LoadAlbumViewModel(private val galleryRepository: GalleryRepository) : ViewModel() {
 
     private val albumLiveData = MutableLiveData<HashSet<PhotoAlbum>>()
 
     fun getAlbums() = albumLiveData
 
-    fun loadAlbums(){
+    fun loadAlbums() {
         albumLiveData.postValue(galleryRepository.getAlbums())
     }
 
